refactor: move validateCountryCode to src/common/country-codes.js
Part 2 wired the cdn-logs-report validateCountryCode into llm-error-pages,
which meant llm-error-pages reached across into another audit's utils. Hoist
the normalizer (and its ISO-3166 set + LLMO-7230 ignore/collision lists) into
src/common/country-codes.js — a peer of country-patterns.js — so both the
cdn-logs-report "hits by market" chart and the llm-error-pages 5xx/4xx
opportunity filter import it from a shared, audit-neutral location.

- new src/common/country-codes.js (100% covered by test/common/country-codes.test.js)
- report-utils.js and llm-error-pages/utils.js and agentic-traffic-mapper.js
  now import from common
- validateCountryCode tests moved out of report-utils.test.js into the new
  dedicated test; llm-error-pages handler.test esmock paths retargeted

No behavior change.
